refactor: modularize the whole dang thing
