+++ constants/constants.go
@@ -1,0 +1,6 @@
+package constants
+
+const BOX_SIZE int = 9
+const MIDPOINT int = 8
+
+const CARD_VIBES int = 10
