+++ deck/deck.go
@@ -1,0 +1,44 @@
+package deck
+
+import "math/rand"
+
+type Deck struct {
+	cards []int
+}
+
+func NewDeck() *Deck {
+	d := Deck{}
+	d.cards = make([]int, 52, 52)
+	count := 0
+	for i := range 13 {
+		for range 4 {
+			d.cards[count] = i + 2
+			count++
+		}
+	}
+
+	d.Shuffle()
+
+	return &d
+}
+
+func (d *Deck) Shuffle() {
+	for i := range d.cards {
+		j := rand.Intn(i + 1)
+		d.cards[i], d.cards[j] = d.cards[j], d.cards[i]
+	}
+}
+
+func (d *Deck) Empty() bool {
+	return len(d.cards) == 0
+}
+
+func (d *Deck) Deal() (int, bool) {
+	if len(d.cards) == 0 {
+		return 0, false
+	}
+	deadNum := d.cards[0]
+	d.cards = d.cards[1:]
+
+	return deadNum, true
+}
