+++ helpers/helpers.go
@@ -1,0 +1,9 @@
+package helpers
+
+func Abs(x int) int {
+	if x < 0 {
+		return -x
+	}
+
+	return x
+}
